Milestone 46: a heap that stays fast, and one kstring for both arches
kheap64.c is a rewrite rather than a port, because the thing worth
porting was the bug. kheap.c keeps one list of every block, used and
free, and kmalloc walks it from the start: N allocations cost N squared.
Milestone 43 measured that as Wine's syscall rate collapsing 21x during
startup. Carrying it to the architecture meant to run a browser was not
defensible.

Two structures instead: blocks doubly linked in address order, so kfree
coalesces with its neighbours in O(1) instead of rescanning the heap;
and a free list, which is the only thing kmalloc walks. Free-list
pointers live in the payload of the free block, so they cost nothing
while a block is in use - the only reason the minimum allocation is 16
bytes.

The test prints the measurement rather than the claim: 1000 steps for
1000 allocations, where quadratic would be ~500000. And after a thousand
alloc/free pairs plus a 128KB allocation that grew the heap, the heap
collapses to a single free block with nothing used - no fragmentation
left. Coalescing is checked from outside, by requiring the free-list
length to return to what it was; with the merges removed it fails and
the count goes 1 -> 1006.

No console in the 64-bit tree, so exhaustion returns NULL and shows in
the counters. The 32-bit version prints, which is part of what made it
non-portable.

kstring.c is now shared by both builds rather than forked. It needed one
change: size parameters are size_t, not uint32_t. Same type on i386, so
the 32-bit build is unaffected; on x86-64 a uint32_t length would have
capped every copy in the kernel at 4GB. kdiv64 turned out not to be
i386-only - it wraps a 32-bit divl, legal in long mode - so it is
compiled and checked rather than excluded.

Verified on the 32-bit side, since kstring is used tree-wide: clean make
with zero warnings under -Wall -Wextra, make test passes (64 fs + 52 wm
checks), make test-qemu passes all 17 transcript assertions.

64-bit boot test: 79 assertions, 0 failures, up from 56.

+++ kernel/kstring.c
@@ -1,52 +1,52 @@
 #include "kstring.h"
 
-void kmemcpy(void* dst, const void* src, uint32_t n) {
+void kmemcpy(void* dst, const void* src, size_t n) {
     uint8_t* d = (uint8_t*)dst;
     const uint8_t* s = (const uint8_t*)src;
-    for (uint32_t i = 0; i < n; i++) d[i] = s[i];
+    for (size_t i = 0; i < n; i++) d[i] = s[i];
 }
 
-void kmemmove(void* dst, const void* src, uint32_t n) {
+void kmemmove(void* dst, const void* src, size_t n) {
     uint8_t* d = (uint8_t*)dst;
     const uint8_t* s = (const uint8_t*)src;
     if (d == s || n == 0) return;
     if (d < s) {
-        for (uint32_t i = 0; i < n; i++) d[i] = s[i];
+        for (size_t i = 0; i < n; i++) d[i] = s[i];
     } else {
-        for (uint32_t i = n; i > 0; i--) d[i - 1] = s[i - 1];
+        for (size_t i = n; i > 0; i--) d[i - 1] = s[i - 1];
     }
 }
 
-void kmemset(void* dst, uint8_t value, uint32_t n) {
+void kmemset(void* dst, uint8_t value, size_t n) {
     uint8_t* d = (uint8_t*)dst;
-    for (uint32_t i = 0; i < n; i++) d[i] = value;
+    for (size_t i = 0; i < n; i++) d[i] = value;
 }
 
-int kmemcmp(const void* a, const void* b, uint32_t n) {
+int kmemcmp(const void* a, const void* b, size_t n) {
     const uint8_t* x = (const uint8_t*)a;
     const uint8_t* y = (const uint8_t*)b;
-    for (uint32_t i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if (x[i] != y[i]) return (int)x[i] - (int)y[i];
     }
     return 0;
 }
 
-void* kmemchr(const void* s, uint8_t c, uint32_t n) {
+void* kmemchr(const void* s, uint8_t c, size_t n) {
     const uint8_t* p = (const uint8_t*)s;
-    for (uint32_t i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         if (p[i] == c) return (void*)(p + i);
     }
     return 0;
 }
 
-uint32_t kstrlen(const char* s) {
-    uint32_t n = 0;
+size_t kstrlen(const char* s) {
+    size_t n = 0;
     while (s[n]) n++;
     return n;
 }
 
-uint32_t kstrnlen(const char* s, uint32_t max) {
-    uint32_t n = 0;
+size_t kstrnlen(const char* s, size_t max) {
+    size_t n = 0;
     while (n < max && s[n]) n++;
     return n;
 }
@@ -56,16 +56,16 @@ int kstrcmp(const char* a, const char* b) {
     return (int)(uint8_t)*a - (int)(uint8_t)*b;
 }
 
-int kstrncmp(const char* a, const char* b, uint32_t n) {
-    for (uint32_t i = 0; i < n; i++) {
+int kstrncmp(const char* a, const char* b, size_t n) {
+    for (size_t i = 0; i < n; i++) {
         if (a[i] != b[i]) return (int)(uint8_t)a[i] - (int)(uint8_t)b[i];
         if (a[i] == '\0') return 0;
     }
     return 0;
 }
 
-int kstrncasecmp(const char* a, const char* b, uint32_t n) {
-    for (uint32_t i = 0; i < n; i++) {
+int kstrncasecmp(const char* a, const char* b, size_t n) {
+    for (size_t i = 0; i < n; i++) {
         char ca = a[i], cb = b[i];
         if (ca >= 'A' && ca <= 'Z') ca = (char)(ca - 'A' + 'a');
         if (cb >= 'A' && cb <= 'Z') cb = (char)(cb - 'A' + 'a');
@@ -90,9 +90,9 @@ char* kstrcpy(char* dst, const char* src) {
     return dst;
 }
 
-char* kstrlcpy(char* dst, const char* src, uint32_t size) {
+char* kstrlcpy(char* dst, const char* src, size_t size) {
     if (size == 0) return dst;
-    uint32_t i = 0;
+    size_t i = 0;
     for (; i + 1 < size && src[i]; i++) dst[i] = src[i];
     dst[i] = '\0';
     return dst;
